ref: extract process termination and cleanup logic into `Killer` module
- Move process termination, signal handling, and directory cleanup methods to `Process::Killer`.
- Simplify `Process` class by delegating termination and cleanup to the new module.
